% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bpvalue.R
\name{bpvalue}
\alias{bpvalue}
\title{Compute Baysian p-value using the response variables' posterior distribution, or  test statistic.}
\usage{
bpvalue(jags.object, obs, sim.obs, fits = NULL, sim.fits = NULL,
  plots = T, store.output = F, path = NULL, method = c("post", "ssq"))
}
\arguments{
\item{jags.object}{Valid jags output (e.g. as produced by jags in packages jags or R2jags)}

\item{obs}{Observations}

\item{sim.obs}{Observations replicated by model simulation}

\item{fits}{optional. Model fits}

\item{sim.fits}{optional. Model fits replicated by model simulation}

\item{plots}{optional. Should plots be produced?}

\item{store.output}{optional. Should plots be stored?}

\item{method}{\code{method = "ssq"} or \code{method = "post"}. Used to compute Bayesian p-value. See details}
}
\value{
Both methods return Bayesian p-values of model fit and optional plots. See details.
}
\description{
Compute Baysian p-value using the response variables' posterior distribution, or  test statistic.
}
\details{
The function calculates Bayesian p-values in to alternative ways that can be specified via the parameter \code{method}.
\itemize{
 \item{\code{method = "post"}} {produces a p-value as measure of model fit from observations and simulated observations, both for the entire model and at node level. The method checks whether the simulated observations produced by the model are distributed around the actually observed value in such a way that the observed value approximates the median of the simulated observations. The if this is the case, the returned p-value would be at around 0.5., indicating good model fit.}
 \item{\code{method = "ssq"}} {looks at the sums of squares (SSQ) of the residuals between the observed response variable and the fitted response variable and the SSQ of the residuals between their respective modelled replicates. The p-value is calculated as the proportion of SSQ of the replicates in all SSQ that are larger than the SSQ of actual observations and predictions (see Kéry and Schaub 2012).}
}
}
\examples{
Both examples use an adoption of the housemartin state-space-model by Kéry and Schaub (2012).

bpvalue(housemartin.ssm2, obs = "y", sim.obs = "S.y", store.output = F, method = "post")

bpvalue(housemartin.ssm2, obs = "y", sim.obs = "S.y", fits = "logN.est", sim.fits = "S
.logN.est", store.output = F, method = "ssq")
}
\author{
Simon Kapitza \email{simon.kapitza@uranus.uni-freiburg.de}
}
\references{
Kéry, M. and Schaub, M. (2012). Bayesian population analysis using WinBUGS: a hierarchical perspective. Academic Press.
}

