% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitplot.R
\name{fitplot}
\alias{fitplot}
\title{Plots of jags posterior distributions}
\usage{
fitplot(jags.object, obs, fits = NULL, store.plot = F, path = NULL,
  filename = NULL)
}
\arguments{
\item{jags.object}{Valid jags output (e.g. as produced by \code{jags} in packages \code{jags} or \code{R2jags})}

\item{obs}{Observations to be plotted}

\item{fits}{Fits to be plotted against observations. Also model simulations}

\item{path}{target path if plots are stored. If \code{store.output = T} and no path is specified, files will be stored in current working directory.}

\item{filename}{intended output file name as character string without file extension}

\item{store.output}{optional. Should plots be stored?}
}
\value{
Function also returns a list containing the specified posterior distributions for further analysis.
}
\description{
This function plots posterior distributions of fits against observations.
}
\details{
The function can be used to plot posterior distributions produced by \code{jags} against observations. This helps to visualise systematic misfits of the modelled values compared to actual observations and supports decisions regarding further analysis. If \code{obs} are distributions, their means will be used as values against which \code{fits} are plotted. Plotted are 95% confidence intervals and the means of the specified posterior distributions, as well as a 1:1 line for easy comparison of posterior distributions and observed variables.
}
\examples{
Both examples use an adoption of the housemartin state-space-model by Kéry and Schaub (2012).
fitplot(housemartin.ssm2, obs = "y", fits = "logN.est")
fitplot(housemartin.ssm2, obs = "y", fits = "S.logN.est", store.plot = T, path = getwd(), filename = "obsvssims")
}
\author{
Simon Kapitza \email{simon.kapitza@uranus.uni-freiburg.de}
}
\references{
Kéry, M. and Schaub, M. (2012). Bayesian population analysis using WinBUGS: a hierarchical perspective. Academic Press.
}

