% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodefunction.R
\name{parnode}
\alias{parnode}
\title{Extraction of Markov chains at node level}
\usage{
parnode(jags.object, parameter, node = NULL, plots = T, store.output = F,
  path = NULL)
}
\arguments{
\item{jags.object}{Valid jags output (e.g. as produced by \code{jags} in packages \code{jags} or \code{R2jags})}

\item{parameter}{Character string containing name of a monitored parameter}

\item{node}{optional. If \code{parameter} was monitored at \code{node}-level, \code{node} must be specified.}

\item{plots}{optional. Should density- and traceplots of \code{parameter} at \code{node} be created?}

\item{store.output}{optional. Should plots be stored?}

\item{path}{target path if plots are stored. If \code{store.output = T} and \code{path} is not specified, files will be stored in current working directory.}
}
\value{
Function returns a list containing the Markov chains of specified \code{parameter} at specified \code{node} (if \code{parameter} was monitored at \code{node} level) for further analysis.
}
\description{
This function extracts Markov chains of specified parameters at specified nodes and creates optional trace and density plots.
}
\details{
The function can be used to extract Markov chains (e.g. posterior distributions) of all \code{parameters} monitored by \code{jags}. Parameters that were monitored at \code{node}-level (e.g. time series of population), \code{node} must be specified. Other parameters, such as \code{standard deviations}, that are monitored globally, do not require a value for \code{node}. Trace and density plots show the posterior distribution (e.g. all Markov chains) of \code{parameter} at \code{node}.
}
\examples{
Both examples apply the housemartin state-space-model by Kéry and Schaub (2012).
parnode(housemartin.ssm2,"r", 15, plots = T, store.output = T)
parnode(housemartin.ssm2,"logN.est", 7, plots = T, store.output = F)
}
\author{
Simon Kapitza \email{simon.kapitza@uranus.uni-freiburg.de}
}
\references{
Kéry, M. and Schaub, M. (2012). Bayesian population analysis using WinBUGS: a hierarchical perspective. Academic Press.
}

